Add field docstrings, to be used by doc automation
The bootstrap config documentation can easily get out of sync, for
which reason we intend to automatically generate the documentation
based on structure docstrings.

To do so, we're copying over the field descriptions from the
following document:

* https://github.com/canonical/k8s-snap/blob/ee017318d958b4c7ed1ab78d3cf4d8b7e94559c5/docs/src/snap/reference/bootstrap-config-reference.md

A separate PR adds the necessary automation. Check
docs/src/_parts/bootstrap_config.md for a sample of the generated
Markdown document.

* https://github.com/canonical/k8s-snap/pull/711

Note that in order to be picked up, the inline comments must be
placed immediately before the corresponding structure field.

Certain comments may only be intended for developers, just add a
new line to avoid including these comments in the generated
documentation:

```
    type SomeStruct struct {
        // skipped comment

        SomeField int `json:"some-field"`

        // multiline comment
        //
        // of a given field
        OtherField string `json:"other-field"`
    }
```