Adds containerd-base-dir field to BootstrapConfig and *JoinConfig
Currently, for the classic k8s snap, the default locations for the
containerd-related files are:

- /etc/containerd/
- /run/containerd/
- /var/lib/containerd/

These paths can conflict with other containerd installations on the host
(e.g. from docker), meaning that the k8s snap cannot be installed. In a
developer's usecase, we shouldn't require the other services to be
disabled, but we can allow the k8s snap's containerd to be installed in
a different location during bootstrap / node join.