Fix bootstrap config field (#26)
Currently, the BootstrapConfig.KubeControllerManagerClientKey can be
unmarshaled from the kube-controller-manager-client-key json field and
kube-ControllerManager-client-key yaml field, which is inconsistent.

Since this is meant to be an API, old value has to be supported for a
while as well. There is no direct support for handling 2 keys for the
same field, so this employs a workaround for it.